backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-environ==0.11.2
django-storages==1.14.2
boto3==1.34.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
opencv-python-headless==4.8.1.78
mediapipe==0.10.8
pandas==2.1.4
scikit-learn==1.3.2
openpyxl==3.1.2
Pillow==10.1.0
python-decouple==3.8
django-cors-headers==4.3.1
scipy==1.11.4 